mcp_gitlab/requirements.txt
# Core dependencies
python-dotenv
python-gitlab
gitpython
omegaconf

# MCP
mcp

# OpenAI integration
openai
azure-ai-inference

# Agent
agno